requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
appnope==0.1.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
blinker==1.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
chromedriver-binary==113.0.5672.24.0
click==8.1.3
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
dearpygui==1.9.0
decorator==5.1.1
discord==2.2.2
discord-protos==0.0.2
discord.py==2.2.2
discord.py-self==2.0.0
dnspython==2.3.0
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.11.0
firebase-admin==6.1.0
fonttools==4.39.3
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
imageio==2.27.0
importlib-metadata==6.6.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
jsonpickle==3.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numpy==1.24.2
openai==0.27.4
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.1
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
polars==0.16.16
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.0
PyJWT==2.6.0
pymongo==4.3.3
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyvis==0.3.2
pywatchman==1.4.1
requests==2.28.2
rich==13.3.4
rsa==4.9
scipy==1.10.1
selenium==4.8.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
streamlit==1.21.0
sympy==1.11.1
tensorboardX==2.6
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torch-ema==0.3
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
trimesh==3.21.5
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
wsproto==1.2.0
yarg==0.1.9
yarl==1.8.2
zipp==3.15.0